WIP: audit cleanup safe 137건 (clang-tidy 125 + cppcheck 12)
audit_20260519_093431 결과 중 safe 카테고리 일괄 처리. 사용자 결정: needs-review 32건은 v0.1.0 후 별도 cleanup PR.

## clang-tidy 125건 fix (잠재 14 별도 commit ea3f4a3)

### 1차 batch (12 fixes)
- performance-move-const-arg (2): YoloV5 push_back std::move 제거 (POD), RtspDetectorUnit metadata_string += 제거
- performance-no-automatic-move (4): DETECTOR.cpp 람다 4건 const std::string/auto 제거 (NRVO 활성)
- cppcoreguidelines-pro-type-member-init (3): RtspDetectorUnit struct tm tstruct {} (3건)
- performance-inefficient-vector-operation (2): MgenSchedule scaled_roi.reserve, YoloV5 indexArray.reserve
- performance-unnecessary-copy-initialization (1): RtspDetectorUnit ImagePath const std::string& 변환

### 2차 batch (69 fixes) — bugprone-narrowing-conversions
모든 implicit narrowing 에 static_cast 명시. behavior 변경 없음 (assembly 동일).

분류:
- int → float: 픽셀 좌표, duration count, vector<float> push_back
- float → int: 좌표 round 결과, bbox 정수 변환
- long → int: chrono duration count, size_t → int 인덱스
- double → float: 0.5/2.0 literal → 0.5f/2.0f
- unsigned int → int: 포트 번호, container size

수정 파일: AbnormalActionChecker (3), GeometricLogic (8), MgenSchedule (4), InferObject (16), YoloV5 (23), RtspDetectorUnit (1), NetworkManager (2), EngineClient (3), GrpcProtoMaker (1), MgenHungarian (3), SORTTracker (1), FramePreProcessor (2), GrpcProtoMaker (1) = 69

### 3차 batch (44 fixes) — performance-unnecessary-value-param
두 패턴 적용:
- "copied for each invocation": T → const T& (signature 변경, header sync)
- "only copied once": signature 유지, body 안 std::move(param)

수정: AbnormalActionChecker LineIntrusionBase class_name, MgenSchedule SetClassEqualChecker cec, EngineHandlerBase lambda input_data, DETECTOR GetAPICommandFromJson target_url, RtspDetectorBlock/Unit ctor shared_ptr 3건 × 2 (std::move), RtspDetectorUnit BuildNotifyJsonImpl_Analysis sch, EngineLoadBalancer input_q, IOStreamManager Register/UnregisterRtspProxies unit_set, ApiHandler BuildCallback_GET checker/url_builder (4), EngineClient::Init lb (const&), condition (std::move), RtspHandler ctor setting, SioHandler RegistEvent event_binder + Set* (6 std::move), GrpcEventClientBase channel, GrpcEventServerBase Set* (11 std::move), MgenKalman ctor init_object/stateMat, FramePreProcessor Convert frame, SwsContextManager GetContext frame.

## cppcheck 12건 fix

- passedByValue (4): EngineStreamTypes Build meta_data const& (2건), AbnormalActionChecker class_name (clang-tidy 와 동일), DETECTOR.cpp target_url (clang-tidy 와 동일)
- shadowVariable (2): EngineHandlerBase input_data → batch_input (2건)
- constParameter / constVariable (3): YoloV5 GetRect img const&, nms outputLocations const&, target_rknn_input const ref
- variableScope (3): SettingManager current_call_success try 안으로, MgenHungarian starMatrixTemp/ptrColumnEnd 루프 안으로, primeRow/primeCol while 안으로

## skip (false positive)

- cppcheck _ dummy (structured binding): unusedVariable 10 + unassignedVariable 8 = 18건 (의도된 throwaway)
- cppcheck std::string_view passedByValue: ServiceProfile ctor service_name, ISettingManager ctor unit_key_name (string_view by value 가 idiomatic)

## defer

- knownConditionTrueFalse 9건: v4 instrument `t_*_set` always true → ThreadProfiler 마이그레이션 시 자연 정리

## 사용자 결정 보류 (v0.1.0 후 cleanup PR)

clang-tidy:
- bugprone-easily-swappable-parameters 16 (refactor)
- bugprone-branch-clone 4 (의도 vs 버그 검토 필요)
- bugprone-string-literal-with-embedded-nul 1 (의도 vs 버그)

cppcheck:
- useStlAlgorithm 5 (for→std::any_of)
- unreadVariable 4 (의도 검토)
- unusedStructMember 2 (FrameDiskBytes used/capacity metric)

합 32건은 별도 refactor/audit-cleanup PR.

build verify: BUILD OK, warning 0 신규 (기존 MgenTypes TRUE/FALSE 매크로 충돌 2건 무관).

+++ code/AbnormalActions/src/GeometricLogic.cpp
@@ -44,11 +44,11 @@ namespace MGEN::Abnormal
 
 	const double getDistance( const MGEN::InferObject& first_box, const MGEN::InferObject& last_box ) noexcept
 	{
-		const int pre_x = first_box.bbox.x + static_cast<int>( first_box.bbox.w / 2 );
-		const int pre_y = first_box.bbox.y + static_cast<int>( first_box.bbox.h / 2 );
+		const int pre_x = static_cast<int>( first_box.bbox.x ) + static_cast<int>( first_box.bbox.w / 2.0f );
+		const int pre_y = static_cast<int>( first_box.bbox.y ) + static_cast<int>( first_box.bbox.h / 2.0f );
 
-		const int cur_x = last_box.bbox.x + static_cast<int>( last_box.bbox.w / 2 );
-		const int cur_y = last_box.bbox.y + static_cast<int>( last_box.bbox.h / 2 );
+		const int cur_x = static_cast<int>( last_box.bbox.x ) + static_cast<int>( last_box.bbox.w / 2.0f );
+		const int cur_y = static_cast<int>( last_box.bbox.y ) + static_cast<int>( last_box.bbox.h / 2.0f );
 
 		const double x = abs(pre_x - cur_x);
 		const double y = abs(pre_y - cur_y);
